# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "はじめてのswagger",
        "title": "gin-swagger guchitter",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/complaints": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complaints"
                ],
                "summary": "Complaintsを全件取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Complaint"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complaints"
                ],
                "summary": "Complaintを一件登録する",
                "parameters": [
                    {
                        "description": "Complaint",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Complaint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登録したComplaint",
                        "schema": {
                            "$ref": "#/definitions/model.Complaint"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/complaints/between-time": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complaints"
                ],
                "summary": "更新日時がfrom, toの間のComplaintを返す",
                "parameters": [
                    {
                        "type": "string",
                        "description": "2022-11-27 0:00:00",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2022-11-28 0:00:00",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Complaint"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/complaints/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complaints"
                ],
                "summary": "avatarIdで検索したComplaintを1件返す",
                "parameters": [
                    {
                        "type": "string",
                        "description": "アバターID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Complaint"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complaints"
                ],
                "summary": "complaintIdで指定したComplaintを1件削除する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "愚痴ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Complaint": {
            "type": "object",
            "properties": {
                "avatarId": {
                    "type": "string",
                    "example": "1"
                },
                "complaintId": {
                    "type": "integer"
                },
                "complaintText": {
                    "type": "string",
                    "example": "勘弁してくれ!"
                }
            }
        }
    }
}